From: Winter Laite <winterlaite@gmail.com> Date: 2015-03-25 16:13 GMT+01:00 Subject: Localized WollMux plugins and other files - Winter To: Christoph Lutz <chrlutz@gmail.com>
Hi, Christoph,

Thanks for your email. Attached are plugins and a few localization files.
The plugins still need a lot of work on my part. The main functionality for
PDF printing is there, however, in English for the most part.
I just modified the build script on my Linux drive to reflect the new Java
version and program locations.

At first I translated pretty aggressively and broke the plugins, so I'm taking
it slow.

The wollmux-config I've downloaded has a number of Munich-specific files and a
few samples. Is this the standalone?

http://www.wollmux.net/files/wollmux-config-en-11.11a-installer.exe

Kind regards,
Winter